identified bug and worked to repair
